+++ src/geometryXYVxVy/vlasov/ivlasovsolver.hpp
@@ -18,15 +18,13 @@ public:
      * @param[in, out] allfdistribu On input : the initial value of the distribution function.
      *                              On output : the value of the distribution function after solving 
      *                              the Vlasov equation.
-     * @param[in] efield_x The electric field in the x direction computed at all spatial positions. 
-     * @param[in] efield_y The electric field in the y direction computed at all spatial positions. 
+     * @param[in] efield The electric field computed at all spatial positions.
      * @param[in] dt The timestep. 
      *
      * @return The distribution function after solving the Vlasov equation.
      */
     virtual DFieldSpVxVyXY operator()(
             DFieldSpVxVyXY allfdistribu,
-            DConstFieldXY efield_x,
-            DConstFieldXY efield_y,
+            DVectorConstFieldXY efield,
             double dt) const = 0;
 };